services/app/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==2.5.2
pydantic-settings==2.1.0
sqlalchemy==2.0.26
psycopg2-binary==2.9.9
pika==1.3.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
PyJWT==2.8.0
bcrypt==4.0.1 